requirements.txt
tensorflow-gpu == 1.2.1
numpy == 1.15.4+mkl
matplotlib == 3.0.2
pandas == 0.20.3
scipy == 0.19.1
